Fix signed integer overflow
The operation "mRsmpInRear += framesRead;" would overflow
the 32-bit variables within 12 minutes for MSD HAL StreamIn IEC61937
MAT 2.0 framed streams (that generate 61440 bytes per 20 milliseconds).
So, increase these counter (and related) variables to 64-bit data types,
(supports up to 95205 years for this data rate).

Test: enable integer_overflow diagnostics, enable MSD HAL with IEC
      MAT patch, wait for 12 minutes for UBSan crash, and check record
      thread frame counts have exceeded 2^31.

+++ services/audioflinger/Threads.h
@@ -1558,7 +1558,7 @@ public:
         size_t              mRsmpInUnrel;   // unreleased frames remaining from
                                             // most recent getNextBuffer
                                             // for debug only
-        int32_t             mRsmpInFront;   // next available frame
+        int64_t             mRsmpInFront;   // next available frame
                                             // rolling counter that is never cleared
     };
 
@@ -1711,7 +1711,7 @@ private:
             size_t                              mRsmpInFramesOA;// mRsmpInFramesP2 + over-allocation
 
             // rolling index that is never cleared
-            int32_t                             mRsmpInRear;    // last filled frame + 1
+            int64_t                             mRsmpInRear;    // last filled frame + 1
 
             // For dumpsys
             const sp<MemoryDealer>              mReadOnlyHeap;
